--- src/ca/ucalgary/seng300/a2/TimeMessage.java
@@ -1,14 +1,0 @@
-package ca.ucalgary.seng300.a2;
-
-public class TimeMessage {
-	public int time;
-	public String message;
-	
-	public  TimeMessage(String MessageIn, int timeIn ) {
-		
-		message = MessageIn;
-		time = timeIn;
-		
-	}
-
-}
